Kafka integration and multiple topic support added

+++ src/Post/PostService.Application/Services/CategoryService.cs
@@ -7,6 +7,8 @@ using Weblog.Shared.Interfaces;
 using PostService.Application.Interfaces;
 using PostService.Application.DTOs.Category;
 using PostService.Domain.Aggreagates;
+using PostService.Domain.Events.Interface;
+using PostService.Domain.Events;
 
 namespace CategoryService.Application.Services;
 
@@ -15,12 +17,14 @@ public class CategoryService : ICategoryService
     private readonly IUnitOfWork _uow;
     private readonly IMapper _mapper;
     private readonly IRepositoryBase<Guid, Category> _repository;
+    private readonly IEventPublisher _eventPublisher;
 
-    public CategoryService(IUnitOfWork unitOfWork, IMapper mapper)
+    public CategoryService(IUnitOfWork unitOfWork, IMapper mapper, IEventPublisher eventPublisher)
     {
         _mapper = mapper;
         _uow = unitOfWork;
         _repository = unitOfWork.Repository<Guid, Category>();
+        _eventPublisher = eventPublisher;
     }
 
     public async Task<Category> Create(CategoryCreateDTO dto)
@@ -41,7 +45,7 @@ public class CategoryService : ICategoryService
         category.SetTitle(categoryDTO.Title);
 
         _repository.Update(category);
-    
+
         await _uow.CommitAsync();
 
         return category;
@@ -51,11 +55,19 @@ public class CategoryService : ICategoryService
     {
         var category = await _repository.Get(id);
 
+        if (category is null)
+            throw new NullReferenceException(nameof(category));
+
         category.RemoveAssociatedPosts();
 
         _repository.Remove(category);
 
         await _uow.CommitAsync();
+
+        // Publish the CategoryDeletedEvent
+        var categoryDeletedEvent = new CategoryDeletedEvent(id);
+        
+        await _eventPublisher.Publish(categoryDeletedEvent, "category");
     }
 
     public async Task<CategoryReadDTO> Get(Guid id) => _mapper.Map<CategoryReadDTO>(await _repository.Get(id));
